requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=base.txt base.in
#
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.7
    # via requests
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
django==2.2.18
    # via
    #   -r base.in
    #   django-filter
    #   drf-yasg
django-filter==2.2.0
    # via -r base.in
djangorestframework==3.10.3
    # via
    #   -r base.in
    #   drf-yasg
drf-yasg==1.20.0
    # via -r base.in
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.0.2
    # via coreschema
markdown==3.1.1
    # via -r base.in
markupsafe==2.0.1
    # via jinja2
packaging==21.2
    # via drf-yasg
psycopg2==2.8.5
    # via -r base.in
pyparsing==2.4.7
    # via packaging
pytz==2019.3
    # via django
requests==2.26.0
    # via coreapi
ruamel.yaml==0.17.17
    # via drf-yasg
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
sqlparse==0.3.0
    # via django
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.7
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
